Fix episodes without totals (???), Fix theatrical identification
Added a default of 24 episodes to shows without a total listed if the current episode is less than 20 & 56 if the current episode is more than 20. still need to add an option for more than 56 but do not want it to go on too far. Plan to use current episode plus 4.
Attempted to fix theatrical display & description